requirements.txt
# Dependencias generales
numpy==1.26.4
pandas==2.2.1
python-decouple==3.8
SQLAlchemy==1.4.49  # Versión compatible con Airflow
psycopg2-binary==2.9.9  # Conexión a PostgreSQL
matplotlib==3.8.3
seaborn==0.13.2
requests==2.32.3

# Airflow y sus dependencias
apache-airflow==2.10.2
apache-airflow-providers-postgres==5.13.0
apache-airflow-providers-google==10.4.0

# Google Drive API
PyDrive==1.3.1  # Para manejar la API de Google Drive